Use their respective files for users and db
Move users endpoint to their respective file
Move database connector in dependencies
