+++ app/dependencies.py
@@ -1,0 +1,10 @@
+# Dependency
+from app.database import SessionLocal
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
