+++ app/main.py
@@ -1,45 +1,14 @@
 from fastapi import Depends, FastAPI, HTTPException
 from app import api
-from sqlalchemy.orm import Session
 
-from . import crud, models, schemas
-from .database import SessionLocal, engine
+from . import models
+from .database import engine
 
 models.Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
 
 app.include_router(api.api_router)
-
-# Dependency
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-@app.post("/users/", response_model=schemas.CoreUser)
-def create_user(user: schemas.CoreUserCreate, db: Session = Depends(get_db)):
-    # db_user = crud.get_user_by_email(db, email=user.email)
-    # if db_user:
-    #     raise HTTPException(status_code=400, detail="Email already registered")
-    return crud.create_user(db=db, user=user)
-
-
-@app.get("/users/", response_model=list[schemas.CoreUser])
-def read_users(db: Session = Depends(get_db)):
-    users = crud.get_users(db)
-    return users
-
-
-@app.get("/users/{user_id}", response_model=schemas.CoreUser)
-def read_user(user_id: int, db: Session = Depends(get_db)):
-    db_user = crud.get_user_by_id(db, user_id=user_id)
-    if db_user is None:
-        raise HTTPException(status_code=404, detail="User not found")
-    return db_user
 
 
 # @app.post("/users/{user_id}/items/", response_model=schemas.Item)
